+++ src/riffy/metadata/__init__.py
@@ -17,13 +17,16 @@ from .audiomoth import AudioMothMetadata
 from .bext import BextMetadata
 from .guano import GuanoMetadata
 from .info import InfoMetadata
-from .recording import RecordingMetadata, read_metadata
+from .ixml import IXmlMetadata
+from .recording import RecordingMetadata, dump_metadata, read_metadata
 
 __all__ = [
     "AudioMothMetadata",
     "BextMetadata",
     "GuanoMetadata",
     "InfoMetadata",
+    "IXmlMetadata",
     "RecordingMetadata",
+    "dump_metadata",
     "read_metadata",
 ]